Move Tempo.Interpolation to Interpolation, updating tests
