+++ Rhythm/Meter.Structure.swift
@@ -64,7 +64,7 @@ extension Meter {
         
         /// - returns: `Tempo.Interpolation` value containing the given `metricalOffset`.
         public func interpolation(containing metricalOffset: MetricalDuration)
-            -> Tempo.Interpolation
+            -> Interpolation
         {
             return tempi.interpolation(containing: metricalOffset)
         }
