+++ RhythmTests/TempoTests.swift
@@ -20,7 +20,7 @@ class TempoTests: XCTestCase {
     
     func testInterpolationNoChange() {
         
-        let interpolation = Tempo.Interpolation(
+        let interpolation = Interpolation(
             start: Tempo(60),
             end: Tempo(60),
             duration: 4/>4
@@ -37,7 +37,7 @@ class TempoTests: XCTestCase {
     
     func testInterpolationChange() {
         
-        let interpolation = Tempo.Interpolation(
+        let interpolation = Interpolation(
             start: Tempo(60),
             end: Tempo(30),
             duration: 4/>4
@@ -50,7 +50,7 @@ class TempoTests: XCTestCase {
     
     func testIntpolationDifferentTempoSubdivisions() {
         
-        let interpolation = Tempo.Interpolation(
+        let interpolation = Interpolation(
             start: Tempo(60, subdivision: 16),
             end: Tempo(60, subdivision: 8),
             duration: 4/>16
@@ -63,7 +63,7 @@ class TempoTests: XCTestCase {
     
     func testLengthOfInterpolation() {
         
-        let interpolation = Tempo.Interpolation(
+        let interpolation = Interpolation(
             start: Tempo(60),
             end: Tempo(120),
             duration: 4/>4
@@ -71,15 +71,15 @@ class TempoTests: XCTestCase {
     }
     
     func testInterpolationTempoAtMetricalOffset() {
-        let interp = Tempo.Interpolation(start: Tempo(30), end: Tempo(60), duration: 4/>4)
+        let interp = Interpolation(start: Tempo(30), end: Tempo(60), duration: 4/>4)
         XCTAssertEqual(interp.tempo(at: 2/>4), Tempo(45))
     }
     
     func testStratum() {
         
         let interpolations = [
-            Tempo.Interpolation(start: Tempo(60), end: Tempo(30), duration: 8/>4),
-            Tempo.Interpolation(start: Tempo(30), end: Tempo(60), duration: 8/>4)
+            Interpolation(start: Tempo(60), end: Tempo(30), duration: 8/>4),
+            Interpolation(start: Tempo(30), end: Tempo(60), duration: 8/>4)
         ]
         
         // Metrical offsets of each interpolation
